+++ src/api/tagapi.h
@@ -13,5 +13,6 @@ class TagApi : public QObject, public Tag {
    public:
     TagApi *fetch(int id);
     Q_INVOKABLE TagApi *fetchByName(const QString &name, int parentId = 0);
+    Q_INVOKABLE QStringList getParentTagNames();
     static TagApi* fromTag(const Tag& tag);
 };